Prototype for displaying flight paths
